add more unit tests

+++ ring_buffer_lib.h
@@ -4,14 +4,14 @@ typedef struct {
 } bitset_t;
 
 typedef struct {
-    int nr_entries;
+    unsigned long nr_entries;
     size_t entry_size;
     // consider storing this as metadata on the entry itself
     bitset_t committed;
     char * buf;
     int * queue;
-    int consumer_pos;
-    int producer_pos;
+    unsigned long consumer_pos;
+    unsigned long producer_pos;
 } rng_buf_t;
 
 typedef struct {
@@ -25,7 +25,7 @@ void release(rng_buf_t * rng_buf, entry_t * entry);
 entry_t reserve(rng_buf_t * rng_buf);
 void commit(rng_buf_t * rng_buf, entry_t * entry);
 
-void init_ring_buf(rng_buf_t * rng_buf, int nr_entries, size_t entry_size);
+void init_ring_buf(rng_buf_t * rng_buf, unsigned long nr_entries, size_t entry_size);
 void destroy_ring_buf(rng_buf_t * rng_buf);
 // for debugging
 void debug_rb(rng_buf_t * rng_buf);
